Updated hieradata with label value
platform::kubernetes::params::host_labels
in hieradata contains all the labels assigned
for a node but does not contain the values
for that keys.

Some label values are needed in the kubernetes
puppet code, e.g :
('openstack-compute-node' in $host_labels) will
not check if the key openstack-compute-node has
the value enabled or disabled.

This commit fixes the bug by also providing access
to the values in the hieradata.

Partial-bug: 1940042
Change-Id: Iefbcb1cde7db27250d9a3629472810d20f208248
Signed-off-by: Daniel Safta <daniel.safta@windriver.com>