k8s upgrade networking before first controller
When upgrading kubernetes we want to upgrade the containerized
networking components first before upgrading the first controller
to the newer version of kubernetes.

The main reason for this is that newer versions of the containerized
networking components are tested with older versions of kubernetes,
but when new versions of kubernetes come out they are not tested
with older versions of the containerized networking components.

By upgrading the containerized networking components first we stay
on the tested combinations as much as possible.

The new ordering will be to download the docker images, then
upgrade networking, then the first controller, then the second
controller, then upgrade kubelet on each node.

Story: 2008972
Task: 43063
Change-Id: Ia8121e0a23d570d4060a4b9fa715c1891c558268
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>