Change error handling for cert validity check
Before this change the code was doing the cert validity check
and stopping the execution if any of the certificates in the
passed in file were invalid.

This change removes the invalid certificate from the PEM
bundle being installed and allows the remaining valid certs
to be installed. The error message is stored and returned at
the end.

Closes-Bug: 1939917
Change-Id: Ic9c45ac15f081576a71472853e6e00ca0017336e
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>