AIO-SX mgmt interface swap with the same vendor/device-id
If the AIO-SX is part of a subcloud, a management interface is
available to connect with the SystemController. As part of
story 2009112, this task adds the mgmt interface on the update MAC
list if the swapped interface has the same vendor and device-id,
on the same PCI address.

Along with an update of the affected interfaces (associated with the
swapped physical port), the MAC address on the host object is also
updated.

The test was executed on a DC setup, by exchanging the card on the
subcloud and possible side effects monitored. After unlocking the
file "dnsmasq.hosts" was updated with the new MAC address. No alarms
were detected and the system controller was able to put the subcloud
back to "managed", "online", and "in-sync". The procedure of backup
and restore was also executed and no problems were found.

Story: 2009112
Task: 43019

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I86aae0d239217154679c56b7b6a70bd05b9bd53f