Fix alarm clearing after resize of database
When doing a resize of the database using controllerfs-modify,
the alarm "controller-0 Configuration is out-of-date" never clears.
The problem happens because the name of controllerfs is "database" while
the comparison uses "pgsql". Only the database has this difference.
The fix modify only places where this constant reference itself, and
don't affect other repositories.

Closes-Bug: 1940963
Signed-off-by: Thiago Miranda <ThiagoOliveira.Miranda@windriver.com>
Change-Id: I82011a9215d98a805db52b986092d56ceb7236a4