In AIO-SX, allow NIC and FEC swap with other vendor or device-id
In a simplex setup, allow the replacement of NIC and FEC devices
without the host reinstallation. This task handles the necessary
updates on the port, interface, and pci-device databases when an
updated report from sysinv-agent arrives at sysinv-conductor. For
multinode setups the NIC and FEC replacement must use the
host-delete and host-add procedure

For NICs, a database search comparing the incoming report with the
entries based on PCI address. If on a particular address, the port's
vendor or device-id doesn't match, then the port DB entry is marked
as replaced. On the same search, if the port DB entry's PCI address
isn't found on the incoming report, it is marked as unreported. With
both lists created, the current port and associated interface are
deleted from the database, if the interface is with class none and
the used_by list is empty. In this case, the new reported port
remains unprocessed until the operator removes the dependent
interface database. If a port replacement occurs, the node_id is
updated from the inode database.

If the operator do take action, by removing the database prior to
the replacement, or the port stops being reported by the OS, an
alarm will be raised.

For FEC devices, a similar procedure is done to create the replaced
and unreported lists, to delete the unnecessary entries from the DB.

In both cases, the new entries will be created using the current
procedures.

Story: 2009112
Task: 43085

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Icc6c11953060f7b4615c14c434cceee57a164d88