Include 'unknown' sync state in startup audit
This change alters the check at cert-mon startup
to include all non-insync states, rather than just
the out-of-sync. This is to account for the 'unknown'
state.

Also including a small change to ensure audit_thread
can be restarted properly. This was noticed during
the cert-alarm reviews (and is similar to cert-mon)

Tested on lab system with large number of subclouds.

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I2758314b44b8bff405fdb297217e78631c2ec7e4