Add subcloud audit parallelization
The subcloud audit for distributed cloud is reworked to audit subclouds
in parallel, using greenthreads.

Parallelization is introduced by organizing the subclouds under audit in
a PriorityQueue, where the queue 'priority' is a timestamp identifying
when the audit should run. To support this, items are inserted into the
queue by the tuple (next-audit-timestamp, SubcloudAuditData)

The existing periodic task infrastructure is maintained but repurposed
as follows: The audit_sc_cert_task now performs the role of scheduler:
pulling any subclouds due for audit off the queue and scheduling them
for execution via a GreenPool.

A new file, subcloud_audit_queue.py, is added to the module to assist
in organizing data associated with the subcloud audits:
- Subcloud audit data is captured in class SubcloudAuditData
- SubcloudAuditPriorityQueue subclasses PriorityQueue to
  provide a specific API for this use case

Story: 2008960
Task: 43242

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I42ffcde16de2a89856b2220eee034175707aa59a