Fix kube-upgrade-download-images step
When running the kube-upgrade-download-images step, the images needed
for the new k8s version are first downloaded by the
https://opendev.org/starlingx/ansible-playbooks/src/branch/master/playbookconfig/src/playbooks/push_k8s_images.yml
playbook.

The list of images is computed by the
https://opendev.org/starlingx/ansible-playbooks/src/branch/master/playbookconfig/src/playbooks/roles/common/load-images-information/tasks/main.yml#L18

ok: [localhost] => {
    "download_images_list": [
        "k8s.gcr.io/kube-apiserver:v1.19.13",
        "k8s.gcr.io/kube-controller-manager:v1.19.13",
        "k8s.gcr.io/kube-scheduler:v1.19.13",
        "k8s.gcr.io/kube-proxy:v1.19.13",
        "k8s.gcr.io/pause:3.2",
        "k8s.gcr.io/etcd:3.4.3-0",
        "k8s.gcr.io/coredns:1.6.7"
    ]
}
And the etcd and coredns image tags are wrong, they should have been
etcd:3.4.13-0 and coredns:1.7.0. This happens because when running
the playbook, the kubeadm binary still points to the 1.18.1 version
and it gets us the wrong images tags.

The fix for this is to update the bindmounts before running the
'push_k8s_images.yml' playbook in order to have the kubeadm binary
pointing to the newer version of k8s (1.19.13) in our case.

Closes-Bug: 1943438
Change-Id: Id73370fb8a148a3680dcb4f8291fddd42b7a271f
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>