Remove sriov_vfs/reported # of VFs from hieradata
The original fix of https://bugs.launchpad.net/starlingx/+bug/1931458
made the parameter 'num_vfs' of hieradata to be used solely to retain
the user configuration and added a new parameter 'sriov_vfs' to store
the value of actual number of VFs reported from device by sysinv. It was
done in this way to avoid the report overwriting the user configuration
in race conditions.
However, it was noticed during patch removal procedures that once the
hieradata had this parameter added by later FEC configuration, after the
patch was removed the parser started to claim about the unknown parame-
ter 'sriov_vfs'.
Since that parameter was added just to keep the reported number of VFs
stored in hieradata (like before when used the 'num_vfs' parameter) this
change removes such "debugging" parameter to avoid patch removal issues.

Closes-Bug: #1931458
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: Ib37d3b29ec029b98ab12017be8af8e5698d2a5c8