Generate self-signed certificate when https is enabled
Currently the self-signed certificate used as the REST API/GUI
server certificate for the first time HTTPS is enabled is generate
at build time and embedded in the ISO. This will make it expire less
than a year from the date the system is installed.

This change removed the certificate generation at build time, instead
generate it at the first time HTTPS is enabled.

Closes-Bug: 1944438
Signed-off-by: Andy Ning <andy.ning@windriver.com>
Change-Id: Ife274cb1b61a76827e27678760e2c90bfdec90ab