Remove pycrypto from requirements.txt
We seem to have stopped using pycrypto a while back but the requirements
remained in requirements.txt. While leaving this alone does not create
problems with the unit tests, it does create issues for Debian package
installs because python3-crypto is not in the debian archive. It is also
no longer maintained as well. This patch removes it.

Tested by running the unit tests locally for both python2.7 and
python3.9

Story: 2009101
Task: 43660

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I9623985ebcead2965fe176b5ff7212bf3336b688