allow application upload for applications with no plugins
Starting with Kubernetes 1.20, the "kubectl cp" command will actually
error out if the item being copied doesn't exist.  Prior to this it
failed silently.

It turns out that our application upload code was relying on the silent
failure.

In order to make it work for applications without plugins we need to
make it explicit in the code that it's not a fatal error if there are
no overrides.

Change-Id: Ifb70907c84b26bf6c2e19a72a60110a20bcb399b
Closes-Bug: 1948327
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>