Support upgrades with 30GB partition
Add an upgrade script to support installing with a 30GB partition. Most
of the work will be updating the size of the platform-backup partition
in the database and then updating the location of the partitions that
follow it.

If there isn't sufficient room for the last partition we will resize it
to fit. We assume the last partition will be larger than the 20GB
increase to platform-backup.

When we shrink the last partition we may need to shrink some of the host
filesystems. If there is free space in the cgts-vg we will use that, any
remaining space required will come from the backup and docker
filesystems.

We assume all partitions are for the cgts-vg.

Tests:
SX: Upgrade from system with no additional partitions
SX: Upgrade with an additional partition and >20GB free space in cgts-vg
SX: Upgrade with an additional partition and <1GB free space in cgts-vg
DX: Upgrade from host with no additional partitions
DX: Upgrade with an additional partition and >20GB free space in cgts-vg

Story: 2008966
Task: 43980
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
Change-Id: I28aa1b3ae6884c6643c5f2a8763e76db8f855e20