improve kubelet upgrade error handling
When upgrading kubelet there are a couple of while loops with
expiry.  Change them to actually check the time instead of assuming
the loop is instantaneous.

Also, if we take an exception while querying the K8s versions we
need to handle the exception in order to ensure that we properly
set the KUBE_HOST_UPGRADING_KUBELET_FAILED state so that the
kubelet upgrade can be retried.

Partial-bug: 1949515
Change-Id: Ic7df945488386bbad8a73dfe574843113993176b
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>