[PTP dual NIC config] Capabilities in PTP tables
This change adds an extra field named "capabilities" to the PTP tables
for instances (services) and interfaces. The purpose of such fields is
to allow those PTP entities to store some metadata associated to them
but not recognized as a "parameter" (which in such case would be stored
in a corresponding entry at the PTP parameters table).

Test Plan:
PASS: Existing unit tests of the affected PTP tables.

Story: 2009248
Task: 43847
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I6eab9a2736688198906cec6c92612b9990b74529