Removing rpm as a dependency for sysinv
A new version of the python module was released Nov 4
which breaks pip installs in some environments.

The rpm dependency is not directly required by sysinv
and so it has been removed.

The source code changes in this review are only included
to ensure that all zuul targets are invoked.

Closes-Bug: 1949804
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I6786c3510bb7bac5f03adf3a3c3d377a090063ac