automatically add ghcr.io docker registry entries on upgrade
In StarlingX 6 the upstream Docker image for Tiller moved from the
gcr.io registry to a new ghcr.io registry.

Early in the bootstrap sequence the R6 load needs access to this
Tiller image, which means that in the case of air-gapped installations
it needs access to the information about the equivalent private
Docker registry.

On new installs this is not a problem, because the R6 load allows
the end user to specify the private registry information via the
localhost.yml file.

For upgrades from R5 however, things are a bit trickier since in R5
we actively prevent the end user from adding new service parameters
for Docker registries that R5 doesn't know about already.

As a workaround, for setups involving multiple controller nodes (i.e.
AIO-DX and multi-node configurations) when migrating the override
information to R6 we will add new entries for the ghcr.io registry
that are exactly equivalent to the gcr.io overrides.  As long as
the customer sets up their private Docker registry server appropriately
this will allow the upgrade to R6 with no additional work on their part.

Testing:
PASS: Designer tests, manually running upgrade script with artificial
      test data added to DB.
PASS: Multi-node lab, upgrading from 21.05 to 21.12

Partial-Bug: 1950200
Change-Id: Idb061dcd019479213bb096e36da9cba3590cb903
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>