Handle exceptions in sysinv due to  missing service parameters
When the ghcr docker service parameter is missing sysinv fails to
download images and spams the logs with an exception. This
exception occurs within another except block and hides the true
error.

This fix changes the variable used in an exception message
when an image download from public/private registry fails. If
the exception occurs, pub_img_tag is not assigned and the
UnboundLocalError issue occurs due to the usage of pub_img_tag in
the exception handler.

This fix improves the original fix in
https://opendev.org/starlingx/config/commit/88cacfc5d582a3bb7524a1a5599b289a52ab1f0d
to fix this scenario.

Test Plan:

PASS: Verify that the exceptions do not occur when the patch
for service parameters is not present

Closes-Bug: 1951014
Change-Id: I77db038d2f33fd42b9b8963fc6b31ebb0c425f6c

Change-Id: I720f63878f6672eccbe1a109f762966d73eef154
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>