Fix issue on secure_system.yaml not being updated
The password in the hieradata secure_system.yaml file unexpectedly
did not update with the new password after being the command
'openstack user password set'. This issue appears in IPV6 and IPV4
controllers. The fix gets the right field inside the object
address.

Files changed after this change:
  /opt/platform/puppet/<software-version>/hieradata/secure_system.yaml

Test Plan:

PASS: Verify on a controller with IPV6 enabled if the 'openstack
      user password set' changes the password on the keystone
PASS: Verify file hieradata/secure_system.yaml is updated with
      new password after 'openstack user password set' command
PASS: Verify on a controller with IPV4 enabled if the 'openstack
      user password set' changes the password on the keystone

Closes-bug: 1949983
Change-Id: I7f624625ac20ddcdeceee7eed305666fc0643170
Signed-off-by: Alexandre Horst <alexandre.horst@windriver.com>