Preserve states for functional user and bmc image
Currently, there are two kinds of functional images: user and bmc.
However there is no attribute to distinguish between the two kinds.
If an user image is already applied, applying bmc image would delete
the state of the applied user image.

A new parameter is added to specify a functional BMC image.
The state of last updated user image and BMC image are to be kept.
The order in which the two kinds of BMC images is applied determine
which one is to be kept.
If a BMC image with retimer is applied before a BMC image without
retimer, the states of both images must be kept.
If a BMC image without retimer is applied before a BMC image with
retimer, the state of the BMC image without retimer can be deleted.

Tox unit tests are added for the various scenarios.

Closes-Bug: 1951602

Change-Id: Ifdcbc09d9f14270a57ecc307fbea7517b048351b
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>