Remove sysinv kube-apiserver upgrade params apply
This can be dropped now that a different fix is submitted on the restore
ansible playbooks. The new fix was initially proposed for platform
restore but also fixes the upgrade scenario.

Test Plan: In an AIO-SX, apply oidc-auth kube-apiserver parameters,
perform an upgrade operation and verify these are properly restored and
applied.

PASS: Verify kube-apiserver service parameters are restored after a
System Upgrade

Closes-Bug: 1948140
Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/818136
Signed-off-by: Rafael Camargos <RafaelLucas.Camargos@windriver.com>
Change-Id: I1908e739046d5b215e0f081836fecdfbf8e3a700