[PTP dual NIC config] Review of PTP data model
Same PTP parameter (key:value) can be shared among different instances
or PTP interfaces. Therefore, instead of listing all the UUIDs of the
owners inside the original 'foreign_uuid' column of the table, a new
"bridge" table is created to manage the many-to-many (N:N) relation-
ship between parameter(s) and owner(s). This is also recommended for
better query results.
Another improvement related to PTP parameters management is the intro-
duction of a parent ("abstract") table to serve as the generic owner of
a PTP parameter, according to the diagram below. Given this hierarchy of
tables it is not necessary to implement a logic to decide upon the
actual owner, leaving such resolution for SQLAlchemy to manage.
Finally, there are explicit one-to-many (1:N) relationships between:
- a PTP instance and the host(s) that use that PTP configuration (then
  removing the host ID column from PtpInstances);
- a PTP interface and the interface(s) that are configured to run PTP
  (then removing the interface ID column from PtpInterfaces).

            +------------------------+
            | PtpParameterOwnerships | ("bridge" table)
            +------------------------+
                        |
                        |
+---------------+       |       +--------------------+
|               |N      |      N|                    |
| PtpParameters |---------------| PtpParameterOwners |
|               |               |                    |
+---------------+               +--------------------+
                                           ^
                                          /_\
                                           |
                                  +--------+--------+
                                  |                 |
                        +--------------+       +---------------+
                        |              |       |               |
                        | PtpInstances |<------| PtpInterfaces |
                        |              |       |               |
                        +--------------+       +---------------+
                              1|                      1|
                               |                       |
                              N|                      N|
                           +-------+             +------------+
                           | Hosts |             | Interfaces |
                           +-------+             +------------+

Test Plan:

TODO: New unit tests to cover the associations between parameter and
      owner, between host and instances, and between interfaces and PTP
      interfaces.

Regression:

PASS: Other unit tests for PTP tables still running (creation of
      instances, parameters, listing, get operations, etc.)
PASS: Fresh install and bootstrap with the new tables

Story: 2009248
Task: 44010
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I27ed713084544ca8fab948f8b5dc8163ebaa8ba6