Support upgrades with additional partition types
During the inital change to support upgrades to 30gb
(https://review.opendev.org/c/starlingx/config/+/817967) we assumed
every partition was assigned to the cgts-vg. This commit will add
support for the other types of partitions that may be present on the
system - nova-local and unassigned.

We will prioritize resizing an unassigned partition first. If all
partitions are assigned to a PV we will use the largest partition.

Additionally an issue was found in DX systems with heterogenous
controller configurations. If the rootfs for controller-1 is a different
disk from controller-0 the initial script would fail when it tried to
enumerate the partitions on that disk. We address this by not using the
database to find the rootfs disk. We instead search for the device using
findmnt and lsblk.

Test Plan:
AIO-SX: Test with unassigned and other partitions
AIO-SX: Test with nova-local and cgts-vg partitions (larger nova-local)
AIO-SX: Test with nova-local and cgts-vg partitions (larger cgts-vg)
DX-Std: Test with heterogenous controller disks

Story: 2008966
Task: 44056
Change-Id: Id9c4fccb4c955ff8ea5d0f238b0a3c7b677785b2
Signed-off-by: David Sullivan <david.sullivan@windriver.com>