Align reconfigure_service_endpoints RPC timeout
The target conductor method reconfigure_service_endpoints() has
a potential timeout of 90 seconds.

reconfigure_service_endpoints() needs to wait for the initial
inventory report which could take about a minute as observed
in the failure case reported.  However, this is already at the
default rpcapi timeout value.

This commit aligns the timeout for the rpcapi call to allow for
the target method timeout and processing time buffer so that the
api call is not timed-out before the target is completed.

Test Plan:
  PASSED Perform bootstrap of controller
  PASSED Perform DC System subcloud upgrade bootstrap
  PASSED Force delay in initial inv beyond default rpc timeout

Closes-Bug: 1952233
Signed-off-by: John Kung <john.kung@windriver.com>
Change-Id: I49d3b46e0750c5316353a7f770541be365428eca