Fix controllerconfig service warning on boot
When the controllerconfig service starts the following
warning is emitted in the dmesg:

Standard output type syslog+console is obsolete, automatically
updating to journal+console. Please update your unit file,
and consider removing the setting altogether.

This is due to a change in systemd-246, where syslog is
no longer supported. Change only for Debian so we
avoid a CentOS retest.

Copied the original controllerconfig.service and replaced
the 'syslog+console' with 'journal+console'.

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I5b41005ae53a1a9dd75dc665a217d607fa80c3f1