Revert "Remove sysinv kube-apiserver upgrade params apply"
This reverts commit 67ec634c2846e64e13c6e3a82ce33ce402ec2ca6.

Reason for revert: During tests, it was detected that the 'Apply puppet
apiserver parameters' restore step is broken on standard systems.
Initial investigations showed that the issue only happens on standard
systems and can be related to the hieradata structure, so this should be
reverted until we have a proper fix for it.

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/819834
Change-Id: I99da8bcb28dc21e8b357702e26463f6e0880a223