Fixing variable renaming
On commit [1] the 3rd parameter for _find_ilvg() was renamed to fix a
lint issue, but the usage of that parameter wasn't correctly renamed and
a problem arised since there is a class in the same module with that
name (that was being shadowed before). This commit fixes that and add
some unitests for this method so we don't merge this kind of error in
the future.

[1] https://opendev.org/starlingx/config/commit/5923349485da0ea32042db1c60b0303f7d12e36d

Partial-Bug: #1952400
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I328c3f5976f1903e621ff170a9da1a929983af18