Fix code to support new json format when calling pg dump_stuck
When doing host-lock in storage nodes when ceph cluster is in warning
state, it verifies the state of the storage node. Doing this involves
calling 'pg dump_stuck' and 'osd crush tree' but their json format has
changed and current code doesn't support this format.

The fix was made to support the new format.

Testing performed:
STANDARD+STORAGE: Set ceph-cluster in warning state and successfully
locked storage-1. To put the cluster in warning state I created a pool
having size 3 and set nosizechange to true. This allows pgs_list to not
be empty and test the loop iteration (which calls 'osd crush tree')

Story: 2009074
Task: 44127

Depends-on: https://review.opendev.org/c/starlingx/utilities/+/820528

Signed-off-by: Vinicius Lopes da Silva <vinicius.lopesdasilva@windriver.com>
Change-Id: Ib68b24151b8e75b5abbe073175a486dd117b31a9