Add validation for a non CA certificate
The user is being able to install SSL certificate as SSL CA
certificate using the system-certificate-install. This should
not be permitted by the system.

So this validation blocks the installation of SSL certificate
if it is passing the mode ssl_ca and using a certificate
that is not CA.

Test Plan:
  PASS: Verify SSL certificate is not being installed as SSL CA
  certificate

Closes-bug: 1954302
Signed-off-by: Alexandre Horst <alexandre.horst@windriver.com>
Change-Id: I036bd1898cf608dd75c521e9d7508960096561e2