[PTP dual NIC config] Changes from new data model
Because of redesign of PTP data model done in:
https://review.opendev.org/c/starlingx/config/+/819391 and extended by
https://review.opendev.org/c/starlingx/config/+/820519 and
https://review.opendev.org/c/starlingx/config/+/820961
the API and related code had to be changed in order to support the
new relationships among tables, the updated workflow of configuration
and so on.
Some unnecessary code was removed for bridge tables that don't need it.

Test Plan:

PASS: New and updated test cases for the PTP configuration workflow.
PASS: Fresh install with commands being executed to populate tables.

Regression:

PASS: All unchanged test cases related to PTP.

Story: 2009248
Task: 44208
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: Ib1fa0a044bfc1a4573084f32d8993f4d4cc2539c