[PTP dual NIC config] Deactivate ptp-instance code
The _get_instance_ptp_config() method was incorrectly being run when
ptp was enabled. This resulted in a bad api call which prevents the
system from being unlocked.

This commit adds a return statement to prevent any of the developmental
ptp-instance code from running and resolves the unlock issue.

Story: 2009248
Task: 44146

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: Idaa9e305740bd132fb8c44c5c0875ddeb0c74e2f