Fix AE mgmt interface reconfig
When the mgmt interface is reconfigured from ethernet to a bonded
interface after the first unlocked of controller-0, the mac address
selected for the bonded interface may be changed depending on the new
interface name added to the bond. This would cause failure in unlocking
the controller because maintenance expects to receive commands from
the original host mgmt mac instead of the mac address of the new
interface.

This commit is to take this case into account by keeping the mac address
of the new bonded interface the same as the original host mgmt mac.

Test Plan:
PASS: Reconfigure mgmt interface from ethernet to bonded interface
PASS: Reconfigure mgmt interface from vlan ethernet to vlan over AE

Closes-Bug: 1955406

Change-Id: I749a3691dedae00055e158d91b12456885f51d36
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>