debian: Fix root partition detection for ostree
The /proc/cmdline behaves a bit differently for ostree based
systems. The hard disk is partitioned for an A/B setup so if
the user takes a snapshot of their system and something goes
wrong then they can rollback their system to an earlier state.
As a result the /proc/cmdline is setup differently from a vanilla
machine.

In order to identify the root disk of an ostree based system we
check for ostree_root's disk label, once that is determined we
check the symlink for /dev/disk/by-label for the disk device.

Test Plan
PASS Run unit tests

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I9fb1f25bb974197b301e7edc7497992b4fe255e2