Support puppet's network configuration for Debian
With the support for both CentOS and Debian, the puppet hiera file
generation for network needs to yield different sets of options.
CentOS uses sysconfig file format and Debian uses ifupdown file
format.

Since this change is focused on providing basic network support,
not all configuration variants are supported, for now only physical
and VLAN interfaces are creating a valid config in
/etc/network/interface. Bonded interfaces aren't adding all necessary
parameters needed to be supported (and it requires the addition
of package ifenslave on the distro). Static routes are also working,
but they require the package ifupdown-extras.

The test of the generated puppet configuration was done using ifup
and ifdown commands from the ifupdown package

Test Plan
PASS: the creation of valid physical interface config
PASS: the creation of valid VLAN interface config
PASS: installation on CentOS

Story: 2009101
Task: 44535

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I0094466db3b8b7c94dfb8d014d8cea98d50214a6