Copy hieradata on openstack when upgrading
During the platform upgrade process, we first install the newer version
of the StarlingX system on the standby controller and after that, we
swact and install the other controller. When this process happens, the
puppet plugins are not in place on the newly updated controller since
the stx-openstack app is installed only on the older version of the
system. This fix bypasses that behavior by checking for all plugins that
uses Openstack configuration data, if it is generating hieradata during
a upgrade process, it will copy those configuration from the older
version's hieradata folder.

TEST PLAN
PASS: Installation from scratch on a new SX server
PASS: Upgrade from previous version to current development version on SX
PASS: Upgrade from previous version to current development version on DX

Closes-Bug: #1961114

Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: Id1cbf682013820f1430bfe292448caeddba585e7