Allow stor deletion in AIO-SX
This feature enables stor deletion in AIO-SX when having more than one
stor configured.

This is required when user wants to replace a disk in
a Simplex system and he adds a second disk to backup data from the
current OSD.

User will delete the replaced disk in a situation where
disk is no longer needed.

Story: 2009866
Task: 44552

Tests:
- AIO-SX: Having two stors, successfully deleted one of them
- AIO-SX: Having a single stor, it was unable to delete the stor
- AIO-SX: Having two stors and a single pool having size == 1, it was
unable to delete the stor

Signed-off-by: Vinicius Lopes da Silva <vinicius.lopesdasilva@windriver.com>
Change-Id: I463ab55cae3e96e32e6612a6de373b581d532513