[PTP dual NIC config] Fix parameters from legacy
In the automatic migration of single-instance PTP configuration for
multi-instance support, the value of 'uds_address' parameter missed the
"ptp4l-" prefix like used in the apply of hieradata in case the user has
not provided it.

Also, the 'update-rate' and 'summary-updates' have to be no longer
migrated, since they are now set with fixed values in the phc2sys
command-line options.

Finally, it fixes the creation of associated PTP interfaces for the
legacy instances, where it should be done for every interface in the old
configuration that was set to either 'slave' or 'master' (not 'none').

Test Plan:

PASS: Patching then checked phc2sys-legacy configuration file.

Task: 44567
Story: 2009248
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I388adc3447c520cd02abce05aa7df346a663c46c